Changes Commit

+++ app/src/main/java/com/example/moviebrowser/model/MoviesModel.java
@@ -88,7 +88,7 @@ public class MoviesModel implements Serializable {
         private String title;
         @SerializedName("vote_average")
         @Expose
-        private int voteAverage;
+        private double voteAverage;
         @SerializedName("overview")
         @Expose
         private String overview;
@@ -185,11 +185,11 @@ public class MoviesModel implements Serializable {
             this.title = title;
         }
 
-        public int getVoteAverage() {
+        public double getVoteAverage() {
             return voteAverage;
         }
 
-        public void setVoteAverage(int voteAverage) {
+        public void setVoteAverage(double voteAverage) {
             this.voteAverage = voteAverage;
         }
